refactor token for ecdsa
